refactor(protocol): extract source test modules
- Move protocol unit tests into sibling *_test.rs files
- Keep library and protocol source files focused on runtime definitions
